% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_drivers.R
\name{noaa_stage2}
\alias{noaa_stage2}
\title{NOAA GEFS forecasts with EFI stage 2 processing
Stage2 processing involves the following transforms of the data:
\itemize{
\item Fluxes are standardized to per second rates
\item Variables are renamed to match CF conventions
\item Fluxes and states are interpolated to 1 hour intervals
}}
\usage{
noaa_stage2(start_date)
}
\arguments{
\item{start_date}{forecast start date yyyy-mm-dd format}
}
\description{
NOAA GEFS forecasts with EFI stage 2 processing
Stage2 processing involves the following transforms of the data:
\itemize{
\item Fluxes are standardized to per second rates
\item Variables are renamed to match CF conventions
\item Fluxes and states are interpolated to 1 hour intervals
}
}
