% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_drivers.R
\name{noaa_stage3}
\alias{noaa_stage3}
\title{NOAA GEFS forecasts with EFI stage 3 processing}
\usage{
noaa_stage3()
}
\description{
Stage 3 processing presents a 'nowcast' product by combining the most
recent predictions from each available cycle. Product uses CF variable
names and 1 hr interval
}
